requirements.txt
streamlit==1.32.0
pydeck==0.8.0
pandas==2.2.0
Pillow==10.2.0
